+++ src/index.js
@@ -1,5 +1,5 @@
 import './styles/main.css';
-import { fetchAndRenderData } from './modules/Api.js';
+import { fetchAndRenderData } from './modules/app.js';
 
 // Call the function to fetch and render data
 fetchAndRenderData();
