+++ src/styles/main.css
@@ -95,3 +95,68 @@ h2 {
   color: #5799ef;
   font-size: 1rem;
 }
+
+    /* modal */
+    #myModal {
+      display: none; /* Hidden by default */
+      position: fixed; /* Stay in place */
+      z-index: 1; /* Sit on top */
+      padding-top: 100px; /* Location of the box */
+      left: 0;
+      top: 0;
+      width: 100%; /* Full width */
+      height: 100%; /* Full height */
+      overflow: auto; /* Enable scroll if needed */
+      background-color: rgb(0,0,0); /* Fallback color */
+      background-color: rgba(0,0,0,0.4); /* Black w/ opacity */
+    }
+    
+    /* Modal Content */
+    .modal-content {
+      background-color: #fefefe;
+      margin: auto;
+      padding: 20px;
+      border: 1px solid #888;
+      width: 70%;
+    }
+    
+    /* The Close Button */
+    .close {
+      color: #aaaaaa;
+      float: right;
+      font-size: 28px;
+      font-weight: bold;
+    }
+    
+    .close:hover,
+    .close:focus {
+      color: #000;
+      text-decoration: none;
+      cursor: pointer;
+    }
+
+    .modal-content h1 {
+      text-align: center;
+      margin-bottom: 20px;
+      font-family: 'Gothic A1', sans-serif;
+    }
+
+    .details {
+      display: grid;
+      grid-template-columns: 1fr 1fr;
+      text-align: center;
+      font-family: 'Gothic A1', sans-serif;
+      font-size: 20px;
+      padding-bottom: 20px;
+    }
+
+    .details p {
+      padding-bottom: 15px;
+    }
+
+    #movieImg {
+      width: 20%;
+      display: block;
+      margin: 0px auto;
+      padding-bottom: 20px;
+    }
