+++ src/template.html
@@ -24,6 +24,14 @@
                 <button id="show-more-button">Show More</button>
             </div>
         </div>
+
+        <!--comments modal--->
+
+        <div id="myModal" class="modal">
+
+            <!-- Modal content -->
+       
+          </div>
     </main>
 </body>
 </html>
