+++ webpack.config.js
@@ -17,7 +17,7 @@ module.exports = {
     static: {
       directory: path.resolve(__dirname, 'dist'),
     },
-    port: 8081,
+    port: 8080,
     open: true,
     hot: true,
     compress: true,
